changed Bestiary to Dictionary
